chore: refactor existing http api to new service
