+++ lib/Service/Controller.js
@@ -3,6 +3,7 @@ import ServiceArtnet from './Artnet.js'
 import ServiceBonjourDiscovery from './BonjourDiscovery.js'
 import ServiceElgatoPlugin from './ElgatoPlugin.js'
 import ServiceEmberPlus from './EmberPlus.js'
+import { ServiceHttpApi } from './HttpApi.js'
 import ServiceHttps from './Https.js'
 import ServiceOscListener from './OscListener.js'
 import ServiceOscSender from './OscSender.js'
@@ -34,9 +35,10 @@ import ServiceVideohubPanel from './VideohubPanel.js'
  */
 class ServiceController {
 	/**
-	 * @param {Registry} registry - the application core
+	 * @param {import('../Registry.js').default} registry - the application core
 	 */
 	constructor(registry) {
+		this.httpApi = new ServiceHttpApi(registry, registry.ui.express.apiRouter)
 		this.https = new ServiceHttps(registry, registry.ui.express, registry.io)
 		this.oscSender = new ServiceOscSender(registry)
 		this.oscListener = new ServiceOscListener(registry)
