+++ lib/UI/Express.js
@@ -17,7 +17,7 @@
 
 import Express from 'express'
 import path from 'path'
-import { isPackaged, ParseAlignment, rgb } from '../Resources/Util.js'
+import { isPackaged } from '../Resources/Util.js'
 import cors from 'cors'
 import fs from 'fs'
 import serveZip from 'express-serve-zip'
@@ -62,6 +62,8 @@ class UIExpress extends Express {
 
 		this.registry = registry
 
+		this.apiRouter = Express.Router()
+
 		this.use(cors())
 
 		this.use((req, res, next) => {
@@ -94,161 +96,7 @@ class UIExpress extends Express {
 			}
 		})
 
-		this.options('/press/bank/*', (req, res, next) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-			res.send(200)
-		})
-
-		this.get('^/press/bank/:page([0-9]{1,2})/:bank([0-9]{1,2})', (req, res) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-
-			this.logger.info(`Got HTTP /press/bank/ (trigger) page ${req.params.page} button ${req.params.bank}`)
-
-			const controlId = this.registry.page.getControlIdAtOldBankIndex(req.params.page, req.params.bank)
-			this.registry.controls.pressControl(controlId, true)
-
-			setTimeout(() => {
-				this.logger.info(`Auto releasing HTTP /press/bank/ page ${req.params.page} button ${req.params.bank}`)
-				this.registry.controls.pressControl(controlId, false)
-			}, 20)
-
-			res.send('ok')
-		})
-
-		this.get('^/press/bank/:page([0-9]{1,2})/:bank([0-9]{1,2})/:direction(down|up)', (req, res) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-
-			if (req.params.direction == 'down') {
-				this.logger.info(`Got HTTP /press/bank/ (DOWN) page ${req.params.page} button ${req.params.bank}`)
-
-				const controlId = this.registry.page.getControlIdAtOldBankIndex(req.params.page, req.params.bank)
-				this.registry.controls.pressControl(controlId, true)
-			} else {
-				this.logger.info(`Got HTTP /press/bank/ (UP) page ${req.params.page} button ${req.params.bank}`)
-
-				const controlId = this.registry.page.getControlIdAtOldBankIndex(req.params.page, req.params.bank)
-				this.registry.controls.pressControl(controlId, false)
-			}
-
-			res.send('ok')
-		})
-
-		this.get('^/rescan', (req, res) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-
-			this.logger.info('Got HTTP /rescan')
-			this.registry.surfaces.triggerRefreshDevices().then(
-				() => {
-					res.send('ok')
-				},
-				() => {
-					res.send('fail')
-				}
-			)
-		})
-
-		this.get('^/style/bank/:page([0-9]{1,2})/:bank([0-9]{1,2})', (req, res) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-
-			this.logger.info(`Got HTTP /style/bank ${req.params.page} button ${req.params.bank}`)
-
-			const controlId = this.registry.page.getControlIdAtOldBankIndex(req.params.page, req.params.bank)
-			const control = this.registry.controls.getControl(controlId)
-
-			if (!control || typeof control.styleSetFields !== 'function') {
-				res.status(404)
-				res.send('Not found')
-				return
-			}
-
-			const newFields = {}
-
-			if (req.query.bgcolor) {
-				const value = req.query.bgcolor.replace(/#/, '')
-				const color = rgb(value.substr(0, 2), value.substr(2, 2), value.substr(4, 2), 16)
-				if (color !== false) {
-					newFields.bgcolor = color
-				}
-			}
-
-			if (req.query.color) {
-				const value = req.query.color.replace(/#/, '')
-				const color = rgb(value.substr(0, 2), value.substr(2, 2), value.substr(4, 2), 16)
-				if (color !== false) {
-					newFields.color = color
-				}
-			}
-
-			if (req.query.size) {
-				const value = req.query.size.replace(/pt/i, '')
-				newFields.size = value
-			}
-
-			if (req.query.text || req.query.text === '') {
-				newFields.text = req.query.text
-			}
-
-			if (req.query.png64 || req.query.png64 === '') {
-				if (req.query.png64 === '') {
-					newFields.png64 = null
-				} else if (!req.query.png64.match(/data:.*?image\/png/)) {
-					res.status(400)
-					res.send('png64 must be a base64 encoded png file')
-					return
-				} else {
-					const data = req.query.png64.replace(/^.*base64,/, '')
-					newFields.png64 = data
-				}
-			}
-
-			if (req.query.alignment) {
-				try {
-					const [, , alignment] = ParseAlignment(req.query.alignment)
-					newFields.alignment = alignment
-				} catch (e) {
-					// Ignore
-				}
-			}
-
-			if (req.query.pngalignment) {
-				try {
-					const [, , alignment] = ParseAlignment(req.query.pngalignment)
-					newFields.pngalignment = alignment
-				} catch (e) {
-					// Ignore
-				}
-			}
-
-			if (Object.keys(newFields).length > 0) {
-				control.styleSetFields(newFields)
-			}
-
-			res.send('ok')
-		})
-
-		this.get('^/set/custom-variable/:name', (req, res) => {
-			res.header('Access-Control-Allow-Origin', '*')
-			res.header('Access-Control-Allow-Methods', 'GET,OPTIONS')
-			res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, Content-Length, X-Requested-With')
-
-			this.logger.debug(`Got HTTP /set/custom-variable/ name ${req.params.name} to value ${req.query.value}`)
-			const result = this.registry.instance.variable.custom.setValue(req.params.name, req.query.value)
-			if (result) {
-				res.send(result)
-			} else {
-				res.send('ok')
-			}
-		})
+		this.use(this.apiRouter)
 
 		/**
 		 * We don't want to ship hundreds of loose files, so instead we can serve the webui files from a zip file
